requirements.txt
requests
numpy
geopandas
matplotlib
pandas
contextily
fpdf2
